Build Misc objects from NUMA nodes if we find a complete graph with minimal distance.
The OS-specific code passes an array of objects and a matrix of
distances to the generic topo_setup_misc_level_from_distances()
routine.

The latter checks the distance matrix sanity and then creates
Misc objects using an array of groupids that is filled by one
of the distance strategy callback:
* topo_setup_groups_from_min_distance_clique() groups objects
  if they form complete graphs with minimal distance. If some
  incomplete minimal distance graph is found, no group is created
  there, and the caller may fallback to another strategy.
* We may add some relaxed strategies to form groups even if there
  is no minimal distance clique.
We may also invoke recursively to factorize the Misc objects
created above using an average distance computed from the
underlying node distances.

In case of problem, TOPO_IGNORE_DISTANCES may be used to disable
all this code for now.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/trunk@801 4b44e086-7f34-40ce-a3bd-00e031736276