Backport/adapt trunk commit r2360: * Don't put -libverbs in HWLOC_LIBS   (we don't need to link the core lib against -libverbs).
This was just a cleanup in trunk but it's a real fix in the 1.0 branch.
We had a useless dependency on libibverbs.so from the core library and
all utils programs (caused by the ibverbs checks still being in the
core configure checks and thus propagating up to HWLOC_EMBEDDED_LIBS).

Found out by installing the 1.0.2 debian package which brought libibverbs1 :)



git-svn-id: http://svn.open-mpi.org/svn/hwloc/branches/v1.0@2361 4b44e086-7f34-40ce-a3bd-00e031736276