Rework parsing of object types in hwloc_mask_append_object() (used by hwloc-calc, hwloc-bind, ...) Accept complex types such as L2cache or group2 there.
This is required to accept as input the output of hwloc -H
(such as "socket:1.L2Cache:0.core:0").

This introduces a new function
 int hwloc_obj_type_sscanf(const char *string, hwloc_obj_type_t *typep, unsigned *depthattrp)
that may actually be interesting to add to the public API (if we're sure that
depthattr is the only object attribute that may ever appear in the typename).

We could also have similar functions to convert between such a type + its optional depth
attribute, and an actual object depth in the tree. But I can't find good names for these
(we already have get_type_depth and get_depth_type, but those don't take care of the
 optional depth attribute).

git-svn-id: http://svn.open-mpi.org/svn/hwloc/trunk@3255 4b44e086-7f34-40ce-a3bd-00e031736276