Fix a nasty bug where objects below a Misc would not be added to any level
Rework and simplify the whole logic to do so:
* Don't try to handle I/O and Misc objects while taking objects,
  take everything and filter-them out later.
* Split-out I/O and Misc special cases to simplify things.

The bug only occurs when Misc objects are already present in the tree,
above other "regular" objects, while connecting levels.
Misc objects are currently inserting after load, so after connecting.
The only case where the bug can currently occur is when restricting
the topology after inserting a Misc because we have to reconnect again.

This fix will also matter later when working on network topologies.

(needs backport in v1.2 and v1.3)

git-svn-id: http://svn.open-mpi.org/svn/hwloc/trunk@3654 4b44e086-7f34-40ce-a3bd-00e031736276