Temporarily disable commit r3717 (adding a group exactly above distances matrices)
It breaks 128ia64-17n4s2c linux test topology because we fails to merge
the newly added group with the machine object. The reason is that the
last numanode has no CPU, hence the insert code considers it as not
comparable with the group. We'll fix this by comparing nodeset too.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/trunk@3724 4b44e086-7f34-40ce-a3bd-00e031736276