Add a testcase with cgroup on Magny-Cours
It's interesting because we want to make sure we don't remove
empty NUMA node that are below sockets.

Only cpus are restricted in the cgroup here.
I'll reduce mems later when other problems are fixed.

Thanks to Ake Sandgren for providing the gathe-topology outputs
(and reporting a problem and testing some patches by the way).

git-svn-id: http://svn.open-mpi.org/svn/hwloc/trunk@3765 4b44e086-7f34-40ce-a3bd-00e031736276