Fix distances in presence of cgroups
cgroups cause some objects to be removed after the initial backend load
and before the final logical distances creation. We can't cache objects
in internal distances structures in the meantime because those objects
may disappear in the meantime.
Fortunately, this is very similar to the restrict code: we drop the
array of objects, and recreate it from the array of physical indexes.

We could optimize this by only dropping the array if some objects
disappeared, but the list of distances is usually very small so
the overhead is very small.

Thanks to Ake Sandgren for reporting the problem on the OMPI users
mailing list and testing some patches. His bug actually disappeared
by now thanks to r3764
  Don't remove empty objects if they contain some NUMA nodes
but we still need this new fix for non-NUMA distances.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/trunk@3768 4b44e086-7f34-40ce-a3bd-00e031736276