Add hwloc_get_obj_index_inside_cpuset()
This is basically the reverse of hwloc_get_obj_inside_cpuset_by_depth/type().

For instance, you want to know that your object is the sixth core of the
second NUMA nodes. You can't just use logical_index because it starts with
cores in the first node. And you can't use sibling_rank because they are
intermediate levels between nodes and cores.

It will be useful to things like Bull's hitopo OMPI component address.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/trunk@3896 4b44e086-7f34-40ce-a3bd-00e031736276