Ignore distance matrices in case of multi-node topologies
The current code relies on cpuset for grouping and for finding
the root object where the matrix is attached, this can't work
with multinode because some objects have no cpusets.

Unfortunately, at least one user already reported his interest
in having distances between multiple nodes. We'll see what we
can do about it.
One idea could consist in adding a hostset on the side of cpusets
and nodesets to identify which hosts are included in a object.



git-svn-id: http://svn.open-mpi.org/svn/hwloc/branches/custom@3916 4b44e086-7f34-40ce-a3bd-00e031736276