Before v1.3, inserting distance matrices in the topology can get wrong if we have CPU-less NUMA-nodes because we only look at cpusets (not at nodesets). We end up inserting the topology in the wrong common ancestors, which means the relative_depth can get wrong.
This patch ignores the matrix if the relative_depth isn't <= 0.
This is a sanity check, but it also ensures that we won't
export XML that cannot be imported (relative_depth must be != 0).

Thanks to Chris Yeoh from IBM for reporting the problem.
(http://www.open-mpi.org/community/lists/hwloc-devel/2011/11/2524.php).


git-svn-id: http://svn.open-mpi.org/svn/hwloc/branches/v1.2-ompi@3951 4b44e086-7f34-40ce-a3bd-00e031736276