Replace an assertion with a error messages in the distance code
If we ever find a NULL covering object instead of a proper root
object for a distance matrix, no need to break with an assert,
just ignore the matrix and print an error message.
It should not happen, but this code is tricky, especially when
there are cgroups and ignoring options, so I'd rather workaround
the problem (and still have people report it) instead of just
asserting false.



git-svn-id: http://svn.open-mpi.org/svn/hwloc/trunk@4315 4b44e086-7f34-40ce-a3bd-00e031736276