Add hwloc-ps -t/--threads to display threads
By default, we still only show bound processes.

If -t is given, we check if each process contains some bound threads.
If so, we print the process as well as all its threads.

If -a -t is given, we display all threads of all processes.

Thanks to François Tessier for testing patches.
For the record, this feature is nice with Charm++ which seems to bind
a single thread within each process. hwloc-ps reports nothing
because each entire process is not bound. hwloc-ps -t shows what's
actually going on.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/trunk@4450 4b44e086-7f34-40ce-a3bd-00e031736276