Fix set_custom() and its disable() behavior
There's really nothing to do in the former, setting the root type
to System and setting thissystem to 0 can be done in the discover()
callback. Then we don't need a disable() callback anymore.

Without this, calling set_custom() twice would reset the root type
to Machine when disabling the first set_custom() right after enabling
the second.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/branches/components@4716 4b44e086-7f34-40ce-a3bd-00e031736276