Embed libtool's libltdl at autogen time
We use the "recursive" mode so that we don't need to run the
libltdl subconfigure recursively (it would be required even
when plugins are disabled). This just adds libltdl stuff to
our config.h and makes its main Makefile usable by us.

This mode was suggested by the libtool guys and we ended up
finding a bug in the way the config header is passed to libltdl.
See http://debbugs.gnu.org/cgi/bugreport.cgi?bug=12262
Hopefully, we have two possible workaround (modifying our CPPFLAGS
or fixing LT_CONFIG_H with AC_CONFIG_COMMANDS_PRE) and we
may get a fixed libtool released when this branch gets released.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/branches/components@4765 4b44e086-7f34-40ce-a3bd-00e031736276