Remove -lhwloc and -no-undefined from plugins ldflags
We don't want the former because of
  http://www.open-mpi.org/community/lists/hwloc-devel/2012/09/3241.php
(warning: big thread)

For the record, I added the former so that Darwin could link plugins.
But it's not needed once we drop the latter.

This breaks plugins support on mingw (libtool wants -no-undefined there),
we'll see if we disable plugins or so.



git-svn-id: http://svn.open-mpi.org/svn/hwloc/branches/components@4816 4b44e086-7f34-40ce-a3bd-00e031736276