Add flags to backend and component structures
Better now than letter

This may be useful to specify backend requirements.

For instance a backend that only uses the notify_new_object()
callback may not need the topology to be fully connected
when it runs, while a backend with a discover() callback
may need levels and/or children to be fully connected.
We could try to reconnect things only when really needed.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/branches/components@4820 4b44e086-7f34-40ce-a3bd-00e031736276