linux: make hwloc_linux_find_kernel_nr_cpus() do what its name says
When we find a large-enough case that makes sched_getaffinity() work,
take the nr_cpu that CPU_ALLOC() actually used,
not the one we gave in input (may be shorter).

The current code always starts with nr_cpus = 1+lastbit(completecpuset),
so we may return a too small nr_cpus, but it'll be large enough for
anything we do with the topology.
But that could have broken later when binding with linux hooks during
the x86 discovery (the completecpuset may not be ready yet).

And it's better to have the function do what it name says.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/trunk@4931 4b44e086-7f34-40ce-a3bd-00e031736276