Rework component enabling
Use a single list for enabled backends
(don't manage OS/global backends separately from additional,
 put all of them inside a single list, in enabling order)

(once set_foo() has optionally enabled some components)
load() scans the entire list of components and enables all
possible ones according to the exclude flags that it progressively
builds.

The HWLOC_COMPONENTS environment variable forces the order of
component enabling.
If a,b is given, a is enabled, b is enabled if not conflicting, then
all other ones.
If a,b,stop is given, the process stops after a and b.
The good way to just move one component x to the top priority is therefore
to set the environment variable to x.

Once components have been enabled as backends, we discover using all
OS/global backends first, then others
(so that libpci doesn't work before PUs exist)

In the meantime, we traverse the list of backends, check whether
they were enable as envvar or not, combine that info with the
topology flags and THISSYSTEM envvar, to find out whether thissystem=0 or 1.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/branches/components@4956 4b44e086-7f34-40ce-a3bd-00e031736276