components: remove the topology from the instantiate callback
It was only used so that backend_alloc() puts it in backend->topology,
we now do that in backend_enable() later.

This clarifies the fact that instantiate() should not modify the topology
neither look at it before deciding what to do.
It just fills a backend depending on optional input parameters.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/branches/components@4997 4b44e086-7f34-40ce-a3bd-00e031736276