Allow components to be dynamically loaded as plugins using libltdl
This is disabled by default.
Pass --enable-plugins to enable this support and build all plugins
that support this (xml_libxml2 and libpci for now, because they
have heavy external dependencies).
Pass --enable-plugins=foo,bar to enable support and only build
components foo and bar as plugins.

This is mostly useful for distribution packages.

Plugins are installed in $(libdir)/hwloc by default.
The HWLOC_PLUGINS_PATH variable may be used to override this.

All tests now get such a variable so that they find plugins
in the build tree before make install.
The variable is added to existing test scripts, or passed
through a the tests/wrapper.sh.in TEST_ENVIRONMENT.

src/topology-fake.c is a fake plugin that may print a debug
message for testing (used in utils/test-fake-plugin.sh.in).

contrib/hwloc-valgrind.supp is updated with some new ltdl suppressions.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/trunk@5020 4b44e086-7f34-40ce-a3bd-00e031736276