Fix doc install when standalone but not building the doc
We still need to enter doc/ during make install (if standalone),
even if building doc is disabled.

So always enter doc/ (if standalone) and check in there what to
do depending on whether HWLOC_BUILD_DOXYGEN and/or HWLOC_INSTALL_DOXYGEN
are set.

Ideally, we would only enter doc/ if HWLOC_BUILD_DOXYGEN,
or doing if make install with HWLOC_INSTALL_DOXYGEN,
but that's not easy to do.

This reverts:
r4676:
  Stop protecting things with HWLOC_BUILD_DOXYGEN (almost)
r4686:
  We don't enter doc/ unless standalone, but the reverse isn't true
  (doc can be disabled in standalone). So make sure we only enter
  doc/ when standalone and doc is enabled.

We're back to what we did in v1.5.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/trunk@5045 4b44e086-7f34-40ce-a3bd-00e031736276