Remove hwloc_bitmap_printf_value(), use hwloc_bitmap_asprintf() directly
The former comes from when we didn't have hwloc_bitmap_asprintf().
It's only used in debug macro, convert these to asprintf() and drop it.

The good thing about this is that this removes an internal static inline
(protected by HWLOC_INSIDE_LIBHWLOC) in favor of using an existing public
function.
And it also cleanups dependencies from plugins to the core.



git-svn-id: http://svn.open-mpi.org/svn/hwloc/trunk@5103 4b44e086-7f34-40ce-a3bd-00e031736276