hwloc/gl.h: big rework of the headers
We cannot use internal GL functions since the GL code may be built as a plugin.
What we just need is some helpers to convert between displays (names or port/device indexes)
and os devices.

Just keep that, as inline helpers,
everything else can be reimplemented on top of it.

Update and cleanup the tests/gl.c test program.

Add the helpers to rename.h and to the doc.

git-svn-id: http://svn.open-mpi.org/svn/hwloc/branches/display@5259 4b44e086-7f34-40ce-a3bd-00e031736276